requirements.txt
Adyen==3.0.0
amqp==2.6.1
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
Automat==0.8.0
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
blinker==1.4
boto3==1.14.52
botocore==1.17.52
braintree==4.3.0
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
clarifai==2.6.2
click==7.1.2
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==5.0.0
constantly==15.1.0
cryptography==2.8
cssselect2==0.3.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.0rc1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1
django-appconf==1.0.4
django-cache-url==3.1.2
django-countries==6.1.3
django-crispy-forms==1.9.2
django-extras==0.3
django-filter==2.3.0
django-form-utils==1.0.3
django-google-maps==0.12.2
django-js-asset==1.2.2
django-measurement==3.2.3
django-money==1.1
django-mptt==0.11.0
django-multi-form-view==2.0.1
django-phonenumber-field==5.0.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-redis==4.12.1
django-render-block==0.7
django-storages==1.10
django-templated-email==2.3.0
django-user-settings==0.0.12
django-versatileimagefield==2.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
EasyProcess==0.3
emoji==0.6.0
enmerkar==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.1.1
filelock==3.0.12
freezegun==0.3.15
future==0.18.2
google-streetview==1.2.9
googleapis-common-protos==1.52.0
grpcio==1.32.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==16.10.1
instapy==0.6.10
Jinja2==2.10.1
jmespath==0.10.0
jsonfield==3.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==2.6.0
keyring==18.0.1
kombu==4.6.11
kwconfig==1.1.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
MeaningCloud-python==2.0.0
measurement==3.2.0
more-itertools==4.2.0
mpmath==1.1.0
netifaces==0.10.4
numpy==1.19.2
oauthlib==3.1.0
Pillow==7.2.0
pipenv==11.9.0
plyer==1.4.3
prices==1.0.0
protobuf==3.13.0
psycopg2-binary==2.8.6
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-magic==0.4.18
python-telegram-bot==12.8
python3-openid==3.2.0
pytz==2020.1
PyVirtualDisplay==1.3.2
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
reverse-geocoder==1.5.1
rsa==4.6
s3transfer==0.3.3
scipy==1.5.2
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
ssh-import-id==5.10
sympy==1.6.2
systemd-python==234
text-unidecode==1.3
tinycss2==1.0.2
tornado==6.0.4
tqdm==4.49.0
Twisted==18.9.0
typing==3.7.4.3
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
webdriverdownloader==1.1.0.3
webencodings==0.5.1
zipp==1.0.0
zope.interface==4.7.1
